在主流程中添加回报率的存储

+++ src/main/scala/stock/RateOfReturn.scala
@@ -5,35 +5,35 @@ package stock
   */
 class RateOfReturn {
 
-  private var _code:String = ""//股票代碼
-  private var _name:String = ""//股票名称
-  private var _current_rate:Double = 0.000 //回报率
+  private var _code: String = ""//股票代碼
+  private var _name: String = ""//股票名称
+  private var _current_rate: Float = 0f //回报率
   private var _date: String = "" //回报率的时间段
 
-  def this(code :String,name : String) = {
+  def this(code: String, name: String) = {
     this()
     this._code = code
     this._name = name
-    this._current_rate = 0.000
+    this._current_rate = 0f
   }
 
-  def setRateOfReturn(current_rate:Double): Unit ={
+  def setRateOfReturn(current_rate: Float): Unit ={
     this._current_rate = current_rate
   }
 
-  def getCode:String = {
+  def getCode: String = {
     this._code
   }
 
-  def getName:String = {
+  def getName: String = {
     this._name
   }
 
-  def getRateOfReturn:Double ={
+  def getRateOfReturn: Float ={
     this._current_rate
   }
 
-  def getRateOfReturnDate:String = {
+  def getRateOfReturnDate: String = {
     this._date
   }
